Add files via upload
The purpose of this recipe application is to retrieve and display meal recipes, using the APIs provided by www.themealdb.com/. Throughout this application, you’ll see that I’ve created API calls and classes to retrieve lists of categories and areas (through which you can filter for specific recipes) and their accompanying list of meals. I’ve also created API calls to retrieve recipe instructions for each meal. 

The object file is where I defined the class objects (i.e., category, meal, recipe, area) to make working with the data retrieved from the API easier. The Request File is the file in which the API calls were coded. In the Recipes file, I created a simple console user face for my application. 

You’ll also see that I’ve added a recipes_ui.py file, in which I’ve created a GUI to display the recipes for a selected category, while displaying the recipe instructions on the right-hand side of the interface.